Doesnt work

+++ dev/include/architecture/triangle.h
@@ -4,6 +4,6 @@
 typedef struct triangle
 {
 	int vert[3];
-	float normal[3];
+	int normal[3];
 }triangle;
 #endif
